changed as requested in pr review

+++ user-service/src/main/java/rs/edu/raf/banka1/dtos/BusinessAccountDto.java
@@ -13,8 +13,8 @@ public class BusinessAccountDto {
     private Double balance;
     private Double availableBalance;
     private Long createdByAgentId;
-    private Integer creationDate;
-    private Integer expirationDate;
+    private Long creationDate;
+    private Long expirationDate;
     private String currency;
     private String accountStatus;
 }
